requirements.txt
future
emoji
lxml
wget
gtts
beautifulsoup4
requests
requests_html
python-telegram-bot==13.13
SQLAlchemy==1.4.39
psycopg2-binary
feedparser
PyYAML
faker
hachoir
pykeyboard
pynewtonmath
aiofiles
ffmpeg-python
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp
Pillow
CurrencyConverter
googletrans==4.0.0rc1
google-trans-new
jikanpy
speedtest-cli
coffeehouse
regex
bleach
wikipedia
markdown2
telethon==1.24.0
telegraph
heroku3
spamwatch
alphabet_detector
pyrate-limiter
cachetools
ujson
pretty_errors
pygithub
TgCrypto
pyrogram==1.4.16
PyDictionary
git+https://github.com/Sorrybabe/downloader
youtube_search_python==1.4.9
youtube_search
lyrics_extractor
asyncio
cloudscraper
dateparser
pymongo
dnspython
secureme
apscheduler
emoji-country-flag
countryinfo
html2text
bs4
redis
pySmartDL
bing_image_downloader
search_engine_parser
pytz
jikanpy
tswift
python-arq
envparse
#profanity (Credits to Julia)
better_profanity
textblob 
nudepy
nekos.py
motor==2.4.0
gpytranslate
fontTools
opencv-python-headless==4.5.4.60
odmantic==0.4.0
tzlocal
pendulum
fuzzysearch
img2pdf
multicolorcaptcha
git+https://github.com/python-telegram-bot/ptbcontrib.git
base58
pybase64
pycryptodome
